pyproject.toml
[project]
name = "jindai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asteval>=1.0.8",
    "authlib>=1.6.6",
    "bs4>=0.0.2",
    "celery>=5.6.2",
    "chardet>=5.2.0",
    "flask>=3.1.2",
    "flask-oidc>=2.4.0",
    "flask-restful>=0.3.10",
    "flask-sqlalchemy>=3.1.1",
    "gguf>=0.10.0",
    "hanlp>=2.1.3",
    "ipython>=9.9.0",
    "isort>=7.0.0",
    "jieba3>=1.0.2",
    "llama-cpp-python>=0.3.16",
    "many-stop-words>=0.2.2",
    "nltk>=3.9.2",
    "numpy>=2.2.6",
    "ocrmypdf>=16.13.0",
    "ocrmypdf-paddleocr-remote",
    "onnxruntime-gpu>=1.23.2",
    "opencc-python-reimplemented>=0.1.7",
    "opencv-python-headless>=4.12.0.88",
    "optimum>=2.1.0",
    "optimum-onnx[onnxruntime-gpu]>=0.1.0",
    "pandas>=2.3.3",
    "pgvector>=0.4.2",
    "pillow>=12.1.0",
    "pip>=25.3",
    "psycopg2-binary>=2.9.11",
    "pycryptodome>=3.23.0",
    "pydantic>=2.12.5",
    "pykakasi>=2.3.0",
    "pymupdf>=1.26.7",
    "pypdf2>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "redis>=7.1.0",
    "regex>=2025.11.3",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "setuptools>=80.9.0",
    "spacy>=3.8.11",
    "sqlalchemy>=2.0.45",
    "textrank4zh>=0.3",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "trafilatura>=2.0.0",
    "transformers>=4.57.3",
    "waitress>=3.0.2",
    "xlsxwriter>=3.2.9",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.uv.sources]
ocrmypdf-paddleocr-remote = { git = "https://github.com/zhuth/ocrmypdf-paddleocr-remote" }
